fixes core::locality::Roster bug resulting in an assert on shutdown
The root problem is the less-then function which allowed certain elements
to be embedded inside multiple, disjoint ranges of other elements. This
was designed to allow multiple calls on nearby objects but resulted in
ordering being essentially broken. The current quick fix is sub-optimal,
splitting apart all elements by function when this could result in cache
misses. In reality, this shouldn't be a problem as it is worst when there
less data.


Former-commit-id: 614d35cf5ee830ac49689400bc0567e55802f0e6